requirements.txt
streamlit
pandas
numpy
pathlib
plotly.graph_object